Escalado de datos antes de select_data()
En scraper_flashscore.py:
	- Evito extraer una season que ya extraje. En caso que ya la extraje, la levanto y continuo la extraccion. Esto es util sobretodo cuando hubo cortes en la extraccion y extrajo solo algunas temporadas de la competicion pero no todas.

En scraper_sofifa.py:
	- Agrego notas segun #parameters y #returns
	- Quito extract_all_teams() puesto que no la uso. En su lugar uso extract_teams(), es decir, extraigo por pais y no para todos los paises juntos.
	- Mejoro prueba()

En clean_data.py:
	- prepare_text_columns(): quite el argumento l_col_to_except puesto que no lo usaba. Ahora prepara todas las columnas object salvo que se indique que columnas específicas se quieren preparar mediante el argumento l_cols_to_process.
	- agregue la eliminacion de estadísticas irrelevantes que hacia en construct_data.py

En integrate_data.py:
	- integrate_player_data_in_match: Agregue un numero mínimo de registros para calcular el promedio. Por ejemplo, si tengo el rating para mas de 7 titulares, guardo el promedio. Si es menos de 7, asigno NaN. Esto evita promedios que son basados en 1 o unos pocos jugadores.

En construct_data.py:
	- Solo reordene funciones y movi delete_not_relevant_stats() a clean_data.py.

En main.py:
- collect_initial_data:
	- dejo de levantar dataframes ya extraídos. Ahora lo hago dentro de la extraccion de Flashscore: antes de extraer una season, se fija si ya la extrajo. En caso de que si, levanta la season ya extraída
	- Agregue la extraccion de df_teams de Sofifa
	- arregle el guardado de dataframes de Sofifa
- clean_data:
	- Quite la eliminación de filas y la hago después de construir datos puesto que es información util para variables históricas como historiales, goles y puntos.
	- Agregue la eliminación de estadísticas irrelevantes que hacia en construct_data(). Solo es una reorganización, no tenia sentido eliminar columnas en construct.
	- Reorganizo el código para ordenarlo por accion (e.g. preparación del texto, eliminación de outliers) en vez de por dataframe.
	- Quite el escalado de la variable valor de mercado puesto que escalo datos en clean_data_2().
- construct_data:
	- Movi eliminación de estadísticas irrelevantes a clean_data().
	- Reorganización de código. Puse historial arriba nada mas.
	- Agregue argumento without_h2h para decidir si construir los historiales o no. Es util para main_next_matches.py para evitar construir historiales (puesto que cuando llamó a construct_data() lo hago con un set reducido para agilizar).
- clean_data_2: —> NUEVA FUNCION!
	- Implemento el escaldo de datos. Lo guardo para poder escalar de la misma forma en main_next_matches.py
	- Agregue la eliminación de columnas que hacia en select_data() puesto que es mucho mas coherente hacerlo en clean. Ademas, quiero escalar después de eliminar columnas y filas con mucho NaN pero antes de select().

En find_best_hyper:
- implemento clean_data_2 luego del tag.
- Implemento guardado de dataframe tageado y dataframe cleaned tras el tag.

En main_next_matches.py:
	- Implemento clean_data_2_new() en la que hago el escalado de datos. Para escalar uso el mismo modelo que use en los datos para entrenar el modelo
	- treat_nan_values(): reemplazo NaN por 0 en historiales puesto que el historial puede ser NaN cuando los equipos no se enfrentaron.
	- main(): cuando no quiero correr missing y aun nunca lo corri para el pais, levanto df_integrated de main.py.

+++ p2_data_understanding/collect_initial_data/scraper_flashscore.py
@@ -570,47 +570,55 @@ def extract_data(id_country, country: str, id_competicion, competition: str, is_
         season_year = crawler.extract_season_year()
         print(f" {season_year} ".center(120, "-"))
 
-        # Click en boton "Mostrar mas partidos" (para ver no solo la jornada actual sino todas las jornadas de la season)
-        crawler.click_show_more_matches()
+        # Si ya extraje la season
+        if not check_if_season_already_extracted(ruta_base, competition_form, season_year):
 
-        # Extraigo partidos (items) y sus ids
-        l_ids = crawler.extract_id_matches()
-        print(f"Partidos recolectados de la season {season_year} (e.g. en premier league deberian ser 380): {len(l_ids)}")
-        progress_bar = tqdm(total=len(l_ids), ncols=80)  # Inicializo barra de progreso
+            # Click en boton "Mostrar mas partidos" (para ver no solo la jornada actual sino todas las jornadas de la season)
+            crawler.click_show_more_matches()
 
-        # POR MATCH (c/u identificado con un id)
-        for id_match in l_ids:
+            # Extraigo partidos (items) y sus ids
+            l_ids = crawler.extract_id_matches()
+            print(f"Partidos recolectados de la season {season_year} (e.g. en premier league deberian ser 380): {len(l_ids)}")
+            progress_bar = tqdm(total=len(l_ids), ncols=80)  # Inicializo barra de progreso
 
-            # Ingreso a pagina de informacion del match
-            url_match = f'https://www.flashscore.com/match/{id_match}/#/match-summary'
-            crawler.driver.get(url_match)
+            # POR MATCH (c/u identificado con un id)
+            for id_match in l_ids:
 
-            # Extraigo todos los datos del partido
-            d_row_match, d_row_match_player, d_row_match_odds, d_row_match_teams, d_row_match_coaches, d_row_match_players = crawler.extract_match_data()
-            d_row_match.update({'id_country': id_country, 'id_competition': id_competicion, 'is_cup': is_cup, 'season': season_year})
+                # Ingreso a pagina de informacion del match
+                url_match = f'https://www.flashscore.com/match/{id_match}/#/match-summary'
+                crawler.driver.get(url_match)
 
-            # Guardo datos del partido
-            df_match = pd.concat([df_match, pd.DataFrame(d_row_match, index=[id_match])])
-            df_match_player = pd.concat([df_match_player, pd.DataFrame(d_row_match_player, index=[id_match])])
-            df_match_odds = pd.concat([df_match_odds, pd.DataFrame(d_row_match_odds, index=[id_match])])
+                # Extraigo todos los datos del partido
+                d_row_match, d_row_match_player, d_row_match_odds, d_row_match_teams, d_row_match_coaches, d_row_match_players = crawler.extract_match_data()
+                d_row_match.update({'id_country': id_country, 'id_competition': id_competicion, 'is_cup': is_cup, 'season': season_year})
 
-            # Guardo solo si las filas son nuevas
-            df_teams = check_new_rows(df_teams, d_row_match_teams)
-            df_coaches = check_new_rows(df_coaches, d_row_match_coaches)
-            df_player = check_new_rows(df_player, d_row_match_players)
-            progress_bar.update(1)
+                # Guardo datos del partido
+                df_match = pd.concat([df_match, pd.DataFrame(d_row_match, index=[id_match])])
+                df_match_player = pd.concat([df_match_player, pd.DataFrame(d_row_match_player, index=[id_match])])
+                df_match_odds = pd.concat([df_match_odds, pd.DataFrame(d_row_match_odds, index=[id_match])])
 
-        progress_bar.close()
+                # Guardo solo si las filas son nuevas
+                df_teams = check_new_rows(df_teams, d_row_match_teams)
+                df_coaches = check_new_rows(df_coaches, d_row_match_coaches)
+                df_player = check_new_rows(df_player, d_row_match_players)
+                progress_bar.update(1)
 
-        if export:
-            # Guardo partidos de la season (por seguridad)
-            df_match.to_excel(f'{ruta_base}/per_season/df_match/{competition_form}_{season_year}.xlsx', index=True)
-            df_match_player.to_excel(f'{ruta_base}/per_season/df_match_player/{competition_form}_{season_year}.xlsx', index=True)
-            df_match_odds.to_excel(f'{ruta_base}/per_season/df_match_odds/{competition_form}_{season_year}.xlsx', index=True)
-            df_teams.to_excel(f'{ruta_base}/per_competition/df_teams/{competition_form}.xlsx', index=True)
-            df_coaches.to_excel(f'{ruta_base}/per_competition/df_coaches/{competition_form}.xlsx', index=True)
-            df_player.to_excel(f'{ruta_base}/per_competition/df_player/{competition_form}.xlsx', index=True)
-    
+            progress_bar.close()
+
+            if export:
+                # Guardo partidos de la season (por seguridad)
+                df_match.to_excel(f'{ruta_base}/per_season/df_match/{competition_form}_{season_year}.xlsx', index=True)
+                df_match_player.to_excel(f'{ruta_base}/per_season/df_match_player/{competition_form}_{season_year}.xlsx', index=True)
+                df_match_odds.to_excel(f'{ruta_base}/per_season/df_match_odds/{competition_form}_{season_year}.xlsx', index=True)
+                df_teams.to_excel(f'{ruta_base}/per_competition/df_teams/{competition_form}.xlsx', index=True)
+                df_coaches.to_excel(f'{ruta_base}/per_competition/df_coaches/{competition_form}.xlsx', index=True)
+                df_player.to_excel(f'{ruta_base}/per_competition/df_player/{competition_form}.xlsx', index=True)
+        
+        else:
+            print(f"La season {season_year} de la competicion {competition_form} ya fue extraida, por lo que, se evita su nueva extraccion.")
+            df_match, df_match_player, df_match_odds, df_teams, df_coaches, df_player = read_dataframes_already_extracted(ruta_base, competition_form, season_year)
+            print(f"Match: {df_match.shape}, Match_player: {df_match_player.shape}, Odds: {df_match_odds.shape}, Teams:{df_teams.shape}, Coaches: {df_coaches.shape}, Player: {df_player.shape}")
+
     if export:
         # Guardo partidos de la competition
         df_match.to_excel(f'{ruta_base}/per_competition/df_match/{competition_form}.xlsx', index=True)
@@ -623,6 +631,28 @@ def extract_data(id_country, country: str, id_competicion, competition: str, is_
     # Finalizada la extraccion, cierro el web browser automático
     crawler.driver.close()
     return df_match, df_match_player, df_match_odds, df_teams, df_coaches, df_player
+
+def check_if_season_already_extracted(ruta_base, competition_form, season_year):
+    """
+    Verificacion de que aun no extraje la temporada
+    """
+    try:
+        pd.read_excel(f'{ruta_base}/per_season/df_match/{competition_form}_{season_year}.xlsx', index_col=0)
+        pd.read_excel(f'{ruta_base}/per_season/df_match_player/{competition_form}_{season_year}.xlsx', index_col=0)
+        pd.read_excel(f'{ruta_base}/per_season/df_match_odds/{competition_form}_{season_year}.xlsx', index_col=0)
+        return True
+    except:
+        return False
+
+def read_dataframes_already_extracted(ruta_base, competition_form, season_year):
+    df_match = pd.read_excel(f'{ruta_base}/per_season/df_match/{competition_form}_{season_year}.xlsx', index_col=0)
+    df_match_player = pd.read_excel(f'{ruta_base}/per_season/df_match_player/{competition_form}_{season_year}.xlsx', index_col=0)
+    df_match_odds = pd.read_excel(f'{ruta_base}/per_season/df_match_odds/{competition_form}_{season_year}.xlsx', index_col=0)
+    df_teams = pd.read_excel(f'{ruta_base}/per_competition/df_teams/{competition_form}.xlsx', index_col=0)
+    df_coaches = pd.read_excel(f'{ruta_base}/per_competition/df_coaches/{competition_form}.xlsx', index_col=0)
+    df_player =  pd.read_excel(f'{ruta_base}/per_competition/df_player/{competition_form}.xlsx', index_col=0)
+    return df_match, df_match_player, df_match_odds, df_teams, df_coaches, df_player
+
 
 def check_new_rows(df: pd.DataFrame, d_row: dict, _print: bool = False):
     """
